pull repos when not downloading

+++ run/03_gitrepos.sh
@@ -68,19 +68,26 @@ function download_repo() {
     git_repo="$2"
     git_email="$3"
     git_branch="$4"
-    echo -en "cloning \e[32m$git_url\e[m into \e[33m$git_repo\e[m, email: \e[34m$git_email\e[m"
-    [[ -n "$git_branch" ]] && echo -en ", branch: \e[35m$git_branch\e[m"
-    echo
-    git clone -c color.ui=always --progress --recurse-submodules "$git_url" "$git_repo"
-    git -C "$git_repo" config user.email "$git_email"
-    [[ -n "$git_branch" ]] && git -C "$git_repo" switch "$git_branch" -q
+    if ! [[ -e "$git_repo" ]]; then
+        echo -en "cloning \e[32m$git_url\e[m into \e[33m$git_repo\e[m, email: \e[34m$git_email\e[m"
+        [[ -n "$git_branch" ]] && echo -en ", branch: \e[35m$git_branch\e[m"
+        echo
+        git clone -c color.ui=always --progress --recurse-submodules "$git_url" "$git_repo"
+        git -C "$git_repo" config user.email "$git_email"
+        [[ -n "$git_branch" ]] && git -C "$git_repo" switch "$git_branch" -q
+    else
+        echo -e "pulling \e[32m$git_url\e[m into \e[33m$git_repo\e[m"
+        git -C "$git_repo" pull --ff-only --progress --recurse-submodules
+    fi
 
     # additional operations done ONLY when repo gets downloaded
     case "$git_repo" in
     "/personal/repos/daniele821/dotfiles")
         # set valid branch for script
-        echo -e "\e[1;34msetting ${git_branch} as the valid branch\e[m"
-        echo "$git_branch" >"${git_repo}/.branch"
+        if ! [[ -e "${git_repo}/.branch" ]]; then
+            echo -e "\e[1;34msetting ${git_branch} as the valid branch\e[m"
+            echo "$git_branch" >"${git_repo}/.branch"
+        fi
         ;;
     "/personal/repos/daniele821/nvim-config")
         # link nvim config repo to ~/.config/nvim
@@ -135,12 +142,10 @@ for ((i = 0; i < "${#GIT_DATA[@]}"; i += 4)); do
     git_repo="${GIT_DATA[$((i + 1))]}"
     git_email="${GIT_DATA[$((i + 2))]}"
     git_branch="${GIT_DATA[$((i + 3))]}"
-    if [[ ! -e "$git_repo" ]]; then
-        TMP_FILE="$(mktemp "${TMP_DIR}/XXXXXXXXXXXX.out")"
-        download_repo "$git_url" "$git_repo" "$git_email" "$git_branch" &>"$TMP_FILE" &
-        CLONEPIDS+=("$!")
-        TMP_FILES+=("$TMP_FILE")
-    fi
+    TMP_FILE="$(mktemp "${TMP_DIR}/XXXXXXXXXXXX.out")"
+    download_repo "$git_url" "$git_repo" "$git_email" "$git_branch" &>"$TMP_FILE" &
+    CLONEPIDS+=("$!")
+    TMP_FILES+=("$TMP_FILE")
 done
 cleanup
